Added basic types number types.

+++ suffix_trie.h
@@ -438,7 +438,7 @@ public:
             if (sl.link()) {
                 links_count = 1;
                 total_links_count += 1;
-                high = std::max(high, 1ULL);
+                high = std::max(high, static_cast<size_t>(1));
             }
             else if (sl.full_leaf()) {
                 auto& leaves = sl.full_leaf_ptr()->m_links;
